+++ NEWS.md
@@ -3,6 +3,8 @@
 * New badge in README.
 * Correction of .gitlab-ci.yml with pkg-config.
 * Correction of `rdb()` examples in the doc and README.
+* Simplification of README.
+* New fonctions `rdb_to_xts()` and `rdb_rename_xts()` that were in README.
 
 # rdbnomics 0.6.2